split task file
